qemu: Fix variable sizing issues with 'bandwidth' argument of qemuBlockCommit
The patch moving the code didn't faithfully represent the typecasting
of the 'bandwidth' variable needed to properly convert from the legacy
'unsigned long' argument which resulted in a build failure on 32 bit
systems:

../src/qemu/qemu_block.c: In function ‘qemuBlockCommit’:
../src/qemu/qemu_block.c:3249:23: error: comparison is always false due to limited range of data type [-Werror=type-limits]
 3249 |         if (bandwidth > LLONG_MAX >> 20) {
      |                       ^

Fix it by returning the check into qemuDomainBlockCommit as it's needed
only because of the legacy argument type in the old API and use
'unsigned long long' for qemuBlockCommit.

Fixes: f5a77198bf9

+++ src/qemu/qemu_block.h
@@ -283,7 +283,7 @@ qemuBlockCommit(virDomainObj *vm,
                 virStorageSource *baseSource,
                 virStorageSource *topSource,
                 virStorageSource *top_parent,
-                unsigned long bandwidth,
+                unsigned long long bandwidth,
                 virDomainAsyncJob asyncJob,
                 virTristateBool autofinalize,
                 unsigned int flags);
